feat(website): rename api-keys page to api-key; add link from user menu
The page is singular since a user can only have one key. Also adds an
"API Key" link to the user dropdown menu.
