+++ website/src/components/auth/UserDropdown.astro
@@ -25,6 +25,9 @@ const showSubscriptions = isStaging();
             )
         }
         <li>
+            <a href='/api-key'>API Key</a>
+        </li>
+        <li>
             <a href='/logout'>Logout</a>
         </li>
     </ul>
